+++ inst/examples/ex-contributions.R
@@ -8,7 +8,7 @@ X <- pca(iris, scale = TRUE)
 head(get_contributions(X, margin = 1))
 
 ## Plot contributions
-viz_contributions(X)
+viz_contributions(X, axes = 1)
 
 ## Plot cos2
 viz_cos2(X)